refactor: Moves parsers to individual files
