+++ assets/scripts/libs/enso/templates/parsers/parsers.js
@@ -1,0 +1,5 @@
+
+export * from './textParser.js';
+export * from './attrParser.js';
+export * from './eventParser.js';
+export * from './refParser.js';
