+++ assets/scripts/libs/enso/templates/parsers/utils.js
@@ -1,0 +1,11 @@
+
+export const getName = attr => attr.name.slice(1).toLowerCase();
+export const getBindings = (source, set) => {
+    let bind;
+    while ((bind = bindEx.exec(source)) !== null) {
+        set.add(bind[1]);
+    }
+};
+
+// Matches object property dependencies, i.e. this.<property>:
+export const bindEx = /(?:this\.)(\w+|\d*)/gi;
